backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary>=2.9.9  # Updated for better Python 3.13 compatibility
# Alternative if psycopg2-binary fails: use psycopg (psycopg3) instead
# psycopg[binary]>=3.1.0
django-extensions==3.2.3

# NumPy (install before pandas for better Windows compatibility)
numpy>=1.26.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Multi-tenancy
django-tenant-schemas==1.10.0

# API Documentation
drf-yasg==1.21.7

# Background Tasks & WebSockets
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.2.0

# File Handling
Pillow>=10.2.0  # Updated for better Windows compatibility
django-storages==1.14.2
boto3==1.29.7

# SMS & Communication
twilio==8.10.0
requests==2.31.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel Import/Export
openpyxl==3.1.2
pandas>=2.2.0  # Updated for better Windows compatibility and pre-built wheels

# Utilities
python-dateutil==2.8.2
pytz==2023.3
django-filter==23.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Monitoring
sentry-sdk==1.38.0

